pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tiger-etf"
version = "0.1.0"
description = "TIGER ETF data pipeline & RDB"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.0",
    "xlrd>=2.0",
    "openpyxl>=3.1",
    "tenacity>=8.2",
    # GraphRAG
    "graphrag-toolkit-lexical-graph @ git+https://github.com/awslabs/graphrag-toolkit.git@v3.16.1#subdirectory=lexical-graph",
    "llama-index-readers-file>=0.1",
    "PyMuPDF>=1.24",
    "boto3>=1.34",
    "neo4j>=5.0",
]

[project.scripts]
tiger-etf = "tiger_etf.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
